backend/requirements.txt
bcrypt==4.1.3
fastapi==0.110.1
psycopg2-binary==2.9.11
pydantic==2.12.5
PyJWT==2.11.0
python-dotenv==1.2.1
starlette==0.37.2
uvicorn==0.25.0
